Fix Legend prop mismatch
