+++ apps/nowcasting-app/components/charts/ChartLegend.tsx
@@ -4,60 +4,12 @@ import { InfoIcon, LegendLineGraphIcon } from "../icons/icons";
 import { FC, useEffect } from "react";
 import useGlobalState from "../helpers/globalState";
 import { getRounded4HoursAgoString } from "../helpers/utils";
-
-const Toggle: FC<{
-  onClick: () => void;
-  visible: boolean;
-}> = ({ onClick, visible }) => {
-  return (
-    <label className="inline-flex items-center cursor-pointer px-1 ml-2">
-      <input type="checkbox" checked={visible} onClick={onClick} className={`toggle`}></input>
-    </label>
-  );
-};
-
-const LegendItem: FC<{
-  iconClasses: string;
-  label: string;
-  dashStyle?: "both" | "dashed" | "none";
-  dataKey: string;
-}> = ({ iconClasses, label, dashStyle, dataKey }) => {
-  const [visibleLines, setVisibleLines] = useGlobalState("visibleLines");
-  const isVisible = visibleLines.includes(dataKey);
-
-  const toggleLineVisibility = () => {
-    if (isVisible) {
-      setVisibleLines(visibleLines.filter((line) => line !== dataKey));
-    } else {
-      setVisibleLines([...visibleLines, dataKey]);
-    }
-  };
-
-  return (
-    <div className="flex items-center flex-initial">
-      <LegendLineGraphIcon className={iconClasses} dashStyle={dashStyle} />
-      <button
-        className="inline-flex flex-1 text-left pl-1 max-w-full w-auto dash:w-auto text-xs dash:text-base dash:tracking-wider dash:pb-1"
-        onClick={toggleLineVisibility}
-      >
-        <span
-          className={`block sometimes-bold w-auto uppercase pl-1${
-            isVisible ? " font-extrabold dash:font-semibold" : " text-ocf-gray-700"
-          }`}
-          title={label}
-        >
-          {label}
-        </span>
-      </button>
-      <Toggle onClick={toggleLineVisibility} visible={isVisible} />
-    </div>
-  );
-};
+import LegendItem from "./LegendItem";
 
 type ChartLegendProps = {
   className?: string;
 };
-export const ChartLegend: React.FC<ChartLegendProps> = ({ className }) => {
+export const ChartLegend: FC<ChartLegendProps> = ({ className }) => {
   const [show4hView] = useGlobalState("show4hView");
   const [nHourForecast, setNHourForecast] = useGlobalState("nHourForecast");
